OFS-59: tests for ContributionPlanBundle's mutations
